Fixed a tiny bug with TGAHeader size in tgaimage.h

+++ tgaimage.h
@@ -4,25 +4,37 @@
 #include <vector>
 
 #pragma pack(push, 1)
+// TGA Header Structure -> It is 18 bytes long
 struct TGAHeader {
+  // Image ID: 1 Byte
   std::uint8_t idlength = 0; // The length of the a "string"
                              // located after the header
+  // =====
 
+  // Color Map Type: 1 Byte
   std::uint8_t colormaptype = 0; // Type of color map: 0 = none, 1 = has palette
+  // =====
+
+  // Image Type Code: 1 Byte
   std::uint8_t imagetypecode = 0; // Type of image 0 = none, 1 = indexed,
                                   // 2 = rgb, 3 = gray, +8 = rle packed
+  // =====
 
+  // Color Map Specification: 5 Bytes
   std::uint16_t colormaptypestart = 0; // First color map entry in palette
   std::uint16_t colormaplength = 0;    // Number of colors in palette
-  std::uint16_t colormapdepth = 0;     // Number of bits per palette entry
+  std::uint8_t colormapdepth = 0;      // Number of bits per palette entry
                                        // 15, 16, 24, 32
+  // =====
 
+  // Image Specification: 10 Bytes
   std::uint16_t x_origin = 0;       // Image X origin
   std::uint16_t y_origin = 0;       // Image Y origin
   std::uint16_t width = 0;          // Image width in pixels
   std::uint16_t height = 0;         // Image height in pixels
   std::uint8_t bitsperpixel = 0;    // Images bits per pixel 8, 16, 24, 32
   std::uint8_t imagedescriptor = 0; // Image descriptor bits (vh flip bits)
+  // =====
 };
 #pragma pack(pop)
